pyproject.toml
[tool.poetry]
name = "nftensor-api"
version = "0.1.0"
description = "Bittensor prompter and image generator"
authors =  ["Fimbrethil", "0xcacti"]
license = "MIT"
readme = "README.md"
packages = [{include = "nftensor_api"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
addict = "2.4.0"
aiohttp = "3.8.4"
aiosignal = "1.3.1"
ansible = "8.0.0"
ansible-core = "2.15.0"
ansible-vault = "2.1.0"
anyio = "3.7.0"
appnope = "0.1.3"
argon2-cffi = "21.3.0"
argon2-cffi-bindings = "21.2.0"
arrow = "1.2.3"
asttokens = "2.2.1"
async-timeout = "4.0.2"
attrs = "23.1.0"
backcall = "0.2.0"
backoff = "2.1.0"
base58 = "2.0.1"
beautifulsoup4 = "4.12.2"
bittensor = "4.1.0rc1.post2"
bleach = "6.0.0"
cachetools = "4.2.4"
certifi = "2020.11.8"
cffi = "1.15.1"
chardet = "3.0.4"
charset-normalizer = "3.1.0"
click = "8.1.3"
comm = "0.1.3"
commonmark = "0.9.1"
cryptography = "39.0.0"
cytoolz = "0.12.1"
dataclasses-json = "0.5.7"
debugpy = "1.6.7"
decorator = "5.1.1"
defusedxml = "0.7.1"
ecdsa = "0.18.0"
exceptiongroup = "1.1.1"
executing = "1.2.0"
fastjsonschema = "2.17.1"
filelock = "3.12.0"
fqdn = "1.5.1"
frozenlist = "1.3.3"
fsspec = "2023.5.0"
fuzzywuzzy = "0.18.0"
google-api-core = "1.34.0"
google-api-python-client = "2.7.0"
google-auth = "1.35.0"
google-auth-httplib2 = "0.1.0"
googleapis-common-protos = "1.59.0"
grpcio = "1.42.0"
grpcio-tools = "1.42.0"
httplib2 = "0.22.0"
huggingface-hub = "0.15.1"
hypothesis = "6.47.4"
idna = "2.10"
importlib-metadata = "6.6.0"
importlib-resources = "5.0.7"
ipykernel = "6.23.1"
ipython = "8.14.0"
ipython_genutils = "0.2.0"
ipywidgets = "8.0.6"
isoduration = "20.11.0"
jedi = "0.18.2"
Jinja2 = "3.0.0"
joblib = "1.2.0"
jsonpointer = "2.3"
jsonschema = "4.17.0"
jupyter = "1.0.0"
jupyter-console = "6.6.3"
jupyter-events = "0.6.3"
jupyter-client = "8.2.0"
jupyter-core = "5.3.0"
jupyter-server = "2.6.0"
jupyter-server-terminals = "0.4.4"
jupyterlab-pygments = "0.2.2"
jupyterlab-widgets = "3.0.7"
langchain = "0.0.189"
loguru = "0.6.0"
MarkupSafe = "2.0.1"
marshmallow = "3.19.0"
marshmallow-enum = "1.5.1"
matplotlib-inline = "0.1.6"
miniupnpc = "2.0.2"
mistune = "2.0.5"
more-itertools = "9.1.0"
msgpack = "1.0.4"
msgpack-numpy = "0.4.7.1"
multidict = "6.0.4"
munch = "2.5.0"
mypy-extensions = "1.0.0"
nbclassic = "1.0.0"
nbclient = "0.8.0"
nbconvert = "7.4.0"
nbformat = "5.9.0"
nest-asyncio = "1.5.6"
netaddr = "0.8.0"
nltk = "3.8.1"
notebook = "6.5.4"
notebook-shim = "0.2.3"
numexpr = "2.8.4"
numpy = "1.21.6"
openapi-schema-pydantic = "1.2.4"
overrides = "7.3.1"
packaging = "23.1"
pandas = "1.5.2"
pandocfilters = "1.5.0"
parso = "0.8.3"
password-strength = "0.0.3.post2"
pexpect = "4.8.0"
pickleshare = "0.7.5"
Pillow = "9.5.0"
platformdirs = "3.5.1"
prometheus-client = "0.14.1"
prompt-toolkit = "3.0.38"
protobuf = "3.19.5"
psutil = "5.9.4"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
py = "1.11.0"
py-bip39-bindings = "0.1.11"
py-ed25519-bindings = "1.0.2"
py-ed25519-zebra-bindings = "1.0.1"
py-sr25519-bindings = "0.2.0"
pyasn1 = "0.5.0"
pyasn1-modules = "0.3.0"
pycparser = "2.21"
pycryptodome = "3.11.0"
pydantic = "1.10.8"
Pygments = "2.15.1"
PyNaCl = "1.5.0"
pyparsing = "3.0.9"
pyrsistent = "0.19.3"
python-dateutil = "2.8.2"
python-json-logger = "2.0.7"
python-Levenshtein = "0.12.1"
pytz = "2023.3"
pyzmq = "25.1.0"
qqdm = "0.0.7"
qtconsole = "5.4.3"
QtPy = "2.3.1"
regex = "2023.5.5"
requests = "2.25.0"
resolvelib = "1.0.1"
retry = "0.9.2"
rfc3339-validator = "0.1.4"
rfc3986-validator = "0.1.1"
rich = "12.5.1"
rsa = "4.9"
scalecodec = "1.2.0"
Send2Trash = "1.8.2"
sentencepiece = "0.1.97"
six = "1.16.0"
sniffio = "1.3.0"
sortedcontainers = "2.4.0"
soupsieve = "2.4.1"
SQLAlchemy = "2.0.15"
stack-data = "0.6.2"
substrate-interface = "1.5.0"
tenacity = "8.2.2"
termcolor = "2.1.1"
terminado = "0.17.1"
textwrap3 = "0.9.2"
tinycss2 = "1.2.1"
tokenizers = "0.13.3"
toolz = "0.12.0"
torch = "1.13.1"
tornado = "6.3.2"
tqdm = "4.64.1"
traitlets = "5.9.0"
transformers = "4.29.2"
typing-inspect = "0.9.0"
typing-extensions = "4.6.3"
uri-template = "1.2.0"
uritemplate = "3.0.1"
urllib3 = "1.26.16"
wcwidth = "0.2.6"
webcolors = "1.13"
webencodings = "0.5.1"
websocket-client = "1.5.2"
widgetsnbextension = "4.0.7"
xxhash = "3.2.0"
yarl = "1.9.2"
zipp = "3.15.0"
python-dotenv = "^1.0.0"
web3 = "5.31.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
